chore(testing): Splits e2e tests from @cubejs-backend/testing-shared
