+++ packages/cubejs-bigquery-driver/package.json
@@ -35,7 +35,7 @@
     "ramda": "^0.27.1"
   },
   "devDependencies": {
-    "@cubejs-backend/testing": "^0.29.31",
+    "@cubejs-backend/testing-shared": "^0.29.31",
     "@types/dedent": "^0.7.0",
     "@types/ramda": "^0.27.40",
     "typescript": "~4.1.5"
