+++ packages/cubejs-bigquery-driver/test/BigQueryDriver.test.ts
@@ -1,4 +1,4 @@
-import { DriverTests } from '@cubejs-backend/testing/dist/src/testing';
+import { DriverTests } from '@cubejs-backend/testing-shared';
 
 import { BigQueryDriver } from '../src';
