+++ packages/cubejs-cubestore-driver/test/CubeStoreQuery.test.ts
@@ -1,5 +1,4 @@
-import { CubeStoreDBRunner } from '@cubejs-backend/testing';
-import { createQueryTestCase, QueryTestAbstract } from '@cubejs-backend/testing/dist/src/testing';
+import { createQueryTestCase, CubeStoreDBRunner, QueryTestAbstract } from '@cubejs-backend/testing-shared';
 import { CubeStoreDriver, CubeStoreQuery } from '../src';
 
 class CubeStoreQueryTest extends QueryTestAbstract<CubeStoreDriver> {
