+++ packages/cubejs-postgres-driver/package.json
@@ -38,7 +38,8 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@cubejs-backend/linter": "^0.29.23",
-    "@cubejs-backend/testing": "^0.29.31",
+    "@cubejs-backend/testing-shared": "^0.29.31",
+    "testcontainers": "^8.4.0",
     "typescript": "~4.1.5"
   },
   "publishConfig": {
