+++ packages/cubejs-testing-shared/.gitignore
@@ -1,0 +1,1 @@
+dist
