+++ packages/cubejs-testing-shared/src/index.ts
@@ -1,2 +1,5 @@
 export * from './query-test.abstract';
 export * from './DriverTests';
+export * from './utils';
+
+export * from './db';
