+++ packages/cubejs-testing/birdbox-fixtures/bigquery.yml
@@ -16,7 +16,7 @@ services:
       CUBEJS_REFRESH_WORKER: "false"
     volumes:
       - ./postgresql/single/cube.js:/cube/conf/cube.js
-      - ./postgresql/schema:/cube/conf/schema
+      - ./postgresql/schema/Orders.js:/cube/conf/schema/Orders.js
     ports:
       - "4000:4000"
     restart: always
