+++ packages/cubejs-testing/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@cubejs-backend/testing",
   "version": "0.29.31",
-  "description": "Cube.js Testing Helpers",
+  "description": "Cube.js e2e tests",
   "author": "Cube Dev, Inc.",
   "license": "Apache-2.0",
   "main": "dist/src/index.js",
@@ -46,6 +46,7 @@
     "@cubejs-backend/query-orchestrator": "^0.29.31",
     "@cubejs-backend/schema-compiler": "^0.29.31",
     "@cubejs-backend/shared": "^0.29.29",
+    "@cubejs-backend/testing-shared": "^0.29.31",
     "@cubejs-client/ws-transport": "^0.29.29",
     "@types/dedent": "^0.7.0",
     "dedent": "^0.7.0",
