+++ packages/cubejs-testing/src/index.ts
@@ -1,2 +1,1 @@
 export * from './birdbox';
-export * from './db';
